Remove procedure
