+++ crates/cubecl-core/src/ir/mod.rs
@@ -4,7 +4,6 @@ mod kernel;
 mod local_allocator;
 mod macros;
 mod operation;
-mod procedure;
 mod processing;
 mod scope;
 mod subcube;
@@ -17,7 +16,6 @@ pub use cmma::*;
 pub use kernel::*;
 pub use local_allocator::*;
 pub use operation::*;
-pub use procedure::*;
 pub use scope::*;
 pub use subcube::*;
 pub use synchronization::*;
