--- crates/cubecl-core/src/ir/procedure/mod.rs
@@ -1,13 +1,0 @@
-mod assign;
-mod base;
-mod early_return;
-mod index;
-mod read;
-mod write;
-
-pub use assign::*;
-pub use base::*;
-pub use early_return::*;
-pub use index::*;
-pub use read::*;
-pub use write::*;
