+++ crates/cubecl-spirv/src/instruction.rs
@@ -19,7 +19,6 @@ impl<T: SpirvTarget> SpirvCompiler<T> {
             Operation::Branch(branch) => self.compile_branch(branch),
             Operation::Metadata(meta) => self.compile_meta(meta),
             Operation::Subcube(subcube) => self.compile_subcube(subcube),
-            Operation::Procedure(_) => todo!(),
             Operation::Synchronization(sync) => self.compile_sync(sync),
             Operation::CoopMma(cmma) => self.compile_cmma(cmma),
         }
